[#186] finished refactoring
